pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-knowledge-vault"
version = "0.1.0"
description = "A personal knowledge base that ingests documents, organizes into an Obsidian vault, vectorizes, clusters, and enriches with AI."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "fuertesito91"}]
keywords = ["knowledge-base", "obsidian", "vector-search", "clustering", "AI"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Text Processing :: General",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "chromadb>=0.4",
    "sentence-transformers>=2.2",
    "scikit-learn>=1.3",
    "anthropic>=0.18",
    "pypdf>=3.0",
    "python-docx>=0.8",
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
]

[project.scripts]
pkv = "pkv.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
